Reservation Serailizer separation : Create / Put,Patch
